+++ CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.5](https://github.com/baloise/design-system/compare/v1.14.4...v1.14.5) (2021-09-13)
+
+
+### Bug Fixes
+
+* **testing:** add label support for checkbox & radio ([47c1dbe](https://github.com/baloise/design-system/commit/47c1dbee1359ec3ee59e60cd85489a2ca9bcaab6))
+
+
+
+## 1.14.3 (2021-09-10)
+
+
+
+
+
+## [1.14.4](https://github.com/baloise/design-system/compare/v1.14.3...v1.14.4) (2021-09-13)
+
+
+### Bug Fixes
+
+* **angular:** use event.detail instead of event.target.value ([4709780](https://github.com/baloise/design-system/commit/47097800fc799de3d0c30a43e1a731836d9ed046))
+
+
+
+
+
+## [1.14.3](https://github.com/baloise/design-system/compare/v1.14.2...v1.14.3) (2021-09-10)
+
+
+### Bug Fixes
+
+* **testing:** look for label instead of input for radio & checkbox ([9c26641](https://github.com/baloise/design-system/commit/9c2664126e1a2e0e0f159755bb1d95d0ac1dfdb3))
+
+
+
+
+
 ## [1.14.2](https://github.com/baloise/design-system/compare/v1.14.1...v1.14.2) (2021-09-10)