+++ packages/components-vue/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.5](https://github.com/baloise/design-system/compare/v1.14.4...v1.14.5) (2021-09-13)
+
+**Note:** Version bump only for package @baloise/design-system-components-vue
+
+
+
+
+
+## [1.14.4](https://github.com/baloise/design-system/compare/v1.14.3...v1.14.4) (2021-09-13)
+
+**Note:** Version bump only for package @baloise/design-system-components-vue
+
+
+
+
+
+## [1.14.3](https://github.com/baloise/design-system/compare/v1.14.2...v1.14.3) (2021-09-10)
+
+**Note:** Version bump only for package @baloise/design-system-components-vue
+
+
+
+
+
 ## [1.14.2](https://github.com/baloise/design-system/compare/v1.14.1...v1.14.2) (2021-09-10)
 
 **Note:** Version bump only for package @baloise/design-system-components-vue